GPS code Added
Folder "GPS" added. The working example is called "test" in this GPS directory, and you can see what files I used in the Makefile. Most of these are part of the SBP library that I cloned, but I wrote gps.c and gps.h, so if you get bugs that's where they'll be.